+++ Tests/BluebirdTests.swift
@@ -726,4 +726,33 @@ class BluebirdTests: XCTestCase {
             }
         waitForExpectations(timeout: defaultTimeout, handler: nil)
     }
+
+    // MARK: - Cancel
+
+    func testCancelSync() {
+        let promise = getInt(5).cancel()
+        XCTAssertTrue(promise.state.isCanceled)
+    }
+
+    func testOnCancelQueue() {
+        let exp = expectation(description: "Promise.cancel")
+        let promise: Promise<Int> = Promise { resolve, reject, onCancel in
+            onCancel(.main) { exp.fulfill() }
+        }
+        DispatchQueue.global().asyncAfter(deadline: .now() + 1) {
+            promise.cancel()
+        }
+        waitForExpectations(timeout: defaultTimeout, handler: nil)
+    }
+
+    func testOnCancelMain() {
+        let exp = expectation(description: "Promise.cancel")
+        let promise: Promise<Int> = Promise { resolve, reject, onCancel in
+            onCancel { exp.fulfill() }
+        }
+        DispatchQueue.global().asyncAfter(deadline: .now() + 1) {
+            promise.cancel()
+        }
+        waitForExpectations(timeout: defaultTimeout, handler: nil)
+    }
 }